+++ src/classify.rs
@@ -177,7 +177,6 @@ impl<Rf: Record> ParallelProcessor<Rf> for GroupKmerProcessor {
             self.kmer_length,
             self.smer_length,
             &mut self.buffers,
-            1,
         );
 
         match &self.buffers.syncmers {
@@ -641,7 +640,7 @@ pub(crate) fn classify_seq_kmers(
     min_hits: u64,
     min_fraction: f64,
 ) -> (usize, Classification) {
-    fill_syncmers(seq, hasher, kmer_length, smer_length, buffers, 1);
+    fill_syncmers(seq, hasher, kmer_length, smer_length, buffers);
 
     for h in hits[..num_groups].iter_mut() {
         *h = 0;